+++ extensions/7.3/install_all.sh
@@ -1,1 +1,14 @@
-../core/install_all.sh
+#!/usr/bin/env bash
+
+set -e
+set -ex
+
+apt-get update
+
+for ext in */; do \
+    cd $ext
+    ext_no_slash=${ext%/}
+    echo "***************** Installing $ext_no_slash ******************"
+    ./install.sh
+    cd ..
+done