+++ utils/generate_cron.php
@@ -33,6 +33,6 @@ foreach ($_SERVER as $key => $command) {
 
 if (($found === true) && !file_exists('/usr/local/bin/supercronic')) {
     // Let's check Supercronic is installed (it could be not installed is we are using the slim version...)
-    error_log('Cron is not available in this image. If you are using the thecodingmachine/php "slim" variant, do not forget to add "ARG INSTALL_CRON=1" in your Dockerfile. Check the documentation for more details.');
+    error_log('Cron is not available in this image. If you are using the verbral/php "slim" variant, do not forget to add "ARG INSTALL_CRON=1" in your Dockerfile. Check the documentation for more details.');
     exit(1);
 }