+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ## New features
 
-- thecodingmachine/php image now has a "slim" variant that does not contain any extension but that can be used
-  to [build the extensions very easily](https://github.com/thecodingmachine/docker-images-php/blob/dfdaa984f0fcc3d66a1b9fef5a6643582deb4d0d/README.md#compiling-extensions-in-the-slim-image).
+- verbral/php image now has a "slim" variant that does not contain any extension but that can be used
+  to [build the extensions very easily](https://github.com/verbral/docker-images-php/blob/dfdaa984f0fcc3d66a1b9fef5a6643582deb4d0d/README.md#compiling-extensions-in-the-slim-image).
 
 ## Breaking changes